pyproject.toml
[tool.poetry]
name = "multitone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.1.0"
numpy = "^1.26.2"
colorama = "^0.4.6"
numba = "^0.58.1"
click = "^8.1.7"

[tool.poetry.scripts]
epson = 'multitone.multitone:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
